app_rpt: Add NATIVE_DSP
Most of the DSP code in app_rpt is almost verbatim
identical to the DSP code in dsp.c in Asterisk. However,
it's not a simple migration since a lot of the current
code uses private dsp.c functions, so the DSP usage in
app_rpt needs to be rewritten to use native Asterisk
DSP APIs.

To commence this task, a NATIVE_DSP macro is added so
we can easily toggle between the current way of doing
things (duplicated DSP functions) and the new way we
will do it (using the Asterisk DSP code). There is a
little bit of work left to be done to make this complete,
and then we'll want to test thoroughly before ultimately
removing the non-NATIVE_DSP code as well as the macro itself.