app_rpt: Don't load repeaters if required tech isn't loaded.
Currently, the rpt loader and reloader does not check to
see if the rxchannel for a repeater is valid. This means
that when an rpt_thread is started for it, it will fail
since the required channel driver doesn't exist or isn't
loaded. More importantly, app_rpt as written will terminate
Asterisk if any threads fail and this is highly undesired
and inappropriate behavior.

An additional check is added such that the repeater will
not be loaded at all if the channel tech is not currently
loaded. This does mean that repeaters may fail to load
if the channel driver does not load before app_rpt.