app_rpt: Refactor some off-nominal path code.
Refactors some repetitive code for failing to obtain
psuedo DAHDI channel into a macro.

Also, we now set rpt_vars.deleted to 1 when this happens,
because this is an error which would not be fixed by
a thread restart; in fact, the same error would just
happen again and again, leading to app_rpt killing
Asterisk the way the main thread logic is currently
written.