chan_simpleusb: Fix deadlock on channel creation.
ast_channel_unlock needs to be called when we
create a new channel. Currently, this isn't done
anywhere and this can lead to a deadlock as the
channel lock is being held permanently and never
released.

This can lead to more deadlocks as things that
hold the channel lock will get blocked on this
channel lock.