app_rpt: fix config concatenation
Prevent the node number from getting appended to the IP address, etc.
This issue arose because previously the format string to snprintf was a variable,
and I think that in most cases this variable didn't contain format characters so
the second character was never used. As a result, it's a little ambiguous to me at
the moment as to what the format string should actually be and when the second
variable would get used or not (digitbuf), but for the moment we just ignore the
second variable so that loopback (rpt fun) works properly and hopefully we'll
figure that bit out later.

Fixes #3.