chan_usbradio: Fix bogus dependencies.
chan_usbradio declared dependencies on ossaudio
and usb, neither of which menuselect found satisfied.

The dependency "oss" also fails but "alsa" works.

We didn't have this issue with chan_simpleusb since
it never declared dependencies for menuselect in the
first place.

chan_usbradio has the same dependencies as chan_simpleusb,
and since it appears to use alsa, we now change the
dependencies to just alsa to make menuselect happy.