app_rpt: Fix memory leaks in statpost.
Fixes some issues that weren't caught when this
was originally ported:

Since we never join the thread we create for making
the actual cURL call, we now create it as detached.
Issue #45 I suspect was caused by this, since
creating non-detached threads and non joining would
likely eventually cause pthread_create to fail.

Secondly, on failure, we now free the memory allocated
for the string to avoid a memory leak.

Finally, if creating the thread fails, print out
the error code.

Fixes #45