rpt_telemetry: Fix telemetry serialization.
For some reason in the new (updated to compile on
modern Asterisk) version of app_rpt, telemetry
wasn't being queued properly. This turned out to be
because when a DAHDI channel hung up near the beginning
of a 2nd telemetry being queued, it would improperly
set active_telem to NULL. When this flag is set to NULL,
this indicates that no other telemetry is active, and thus
the 2nd telemetry would go right on ahead and overlay on
top of the original one.

Now we only set to NULL if the active telemetry was in
fact the telemetry for the thread in question, to avoid
this type of improper behavior. This whole part of the code
could still be written to be more robust, but this seems
to fix the issue for now.

Fixes #43