app_rpt: Lock the channel when calling ast_sendtext.
The channel needs to be locked when calling ast_sendtext.
Because it isn't now, occasionally when called, we can get
a thread blocking scenario since another thread might be
doing something with the channel. This causes an assertion
and can cause Asterisk to crash if DO_CRASH is enabled.

We now lock the channel while calling ast_sendtext.

Fixes #47