app_rpt: Add debugging comments for failed text frame transmission.
This really was a doozy of an issue to troubleshoot, with a slight
bit of a causal chain to follow.

The high-level problem that was encountered was that connects to
a system running the new, revamped app_rpt (current code) were causing
the other node to think that we were keyed (transmitting). This turned
out to be because we were transmitting voice frames to the other end,
when we shouldn't have been doing this.

The underlying cause for this (which required a little bit of reverse
engineering Jim Dixon's not-at-all-even-a-little-bit commented code)
was that critical text frames that we needed to see in order to set
newkey and clear the newkeytimer were not being received on time. As
a result, the timer would expire and newkey would be set to 0, which
at a different part of the code made us start passing voice frames
when we shouldn't have been passing those on.

Investigating further, this turned out to be due to an issue with
chan_iax2's jitter buffer handling. Disabling the jitterbuffer in
iax.conf worked around the issue, as it turned out that the frames
were being properly sent, but weren't getting passed up from chan_iax2
into Asterisk when they should have been. Receiving a voice frame would
fix this and allow the queue to be read, but at this point it was too late.

The actual resolution to this is not part of this commit, but a separate
change for chan_iax2 has been made and pushed up for review on Gerrit,
here: https://gerrit.asterisk.org/c/asterisk/+/19712

(Keep in mind, Asterisk is moving away from Gerrit soon, so this link
won't be valid for very much longer). Hopefully it gets merged and it
will be in the commit history though.

This commit is pulled in for Asterisk installs using PhreakScript so
the issue has been fixed. As far as this commit goes, it includes mainly
comments about certain parts of the code that document what they do and
why, thanks to spending hours and hours troubleshooting this issue.

Fixes #46