chan_voter: Fix segfault on unload.
chan_voter would previously segfault when
unloaded, and the backtrace was not very
useful other than suggesting significant
memory corruption. One clue was that if
a usleep was added in the unload for a
sufficient period of time (a few hundred
milliseconds), then no issue would occur.

The fix is to not make the two core threads
detached and to instead join to them unload
so they finish before proceeding with the
unload.

Fixes #21