rpt_manager: Fix assertion if rxchannel disappears.
Exposed by a high-frequency polling AMI application,
if the rxchannel disappears while manager actions are running,
this causes an assertion since we try to lock a channel
which has gotten hung up.

This is prevented by using ast_channel_get_by_name to verify
the channel still exists, before we try to lock it, then unreffing
it afterwards if did. However, I think a better long term fix is
needed, as the channels should probably be ref'ed the whole time
that app_rpt is using them, to prevent this from happening
in possibly other places as well.

Fixes #53