app_rpt: Fix assertions due to duplicate servicing.
Multiple RPT threads can simultaneously perform blocking
operations on channels, which can lead to assertions
(which cause Asterisk to abort in DEVMODE).

Locking the channel prior to such operations results
in noticable artifacts, but holding the RPT lock did not.
To ensure no impact to other locks, an additional lock
has been added around operations that perform these
kinds of blocking operations (e.g. waitfor, sendtext)
that previously led to assertions frequently. In testing,
the assertions seem to have disappeared with no noticable
impact to anything else.

One downside of this approach is that waitfor is blocking,
so it now appears that locks are perpetually held when running
"core show locks". This is fine (it's not a deadlock), but could
be jarring. Certainly, this is not 100% ideal from an architecture
perspective, and there may be better ways to address this within
app_rpt itself, but this ensures that such operations are properly
serialized for now.

Fixes #34
Fixes #58