app_rpt: Don't reload node configs on startup.
Node configs were always previously loaded twice
on startup at about the same time, which is
unnecessary and could cause confusion due to the
load and immediate reload.

This tries to be a little bit more clever about
not reloading configs unnecessarily, although
there are probably more optimizations that
could be done.

Fixes #52