rpt_install: Attempt to update self when run.
If a copy of rpt_install.sh is run from outside
the repo, it will not update itself and potentially
fail to do things that it should, if the script
has been updated.
This adds a very crude mechanism to ensure
the script is up to date by copying it from the repo
when run.

Fixes #41