app_rpt: Minor performance optimizations.
The fact that CLI commands are used under the
hood in app_rpt to communicate with chan_tlb
and chan_echolink is far from optimal, but if
the modules are not loaded, then there's no
point in running them, so we short circuit this
if the module is not currently running. This
now prevents "No such CLI command" type messages
from showing up in the debug logs.

Additionally, it seems that AMI requests can
try to access DAHDI/pseudo, which isn't a real
channel that exists, so there's no point in
traversing the channels container in this case.
A simple check skips the traversal to speed up
response handling.

Fixes #17