res_usbradio: Refactor USB channel driver code.
Much of the code in chan_usbradio and chan_simpleusb
is duplicated between the two channel drivers.

This refactors code that doesn't depend on the channel pivot
into a separate resource module, res_usbradio,
that is now a dependency for both channel drivers,
to help eliminate some of the code duplication.

Closes #118