rpt_manager: Fix DAHDI pseudo channels being ignored.
DAHDI pseudo channels were being skipped because the
name stored in the rpt structure is DAHDI/pseudo,
not the entire channel name (which obviously would not
be known in advance). However, we were using this name
to try to retrieve the channel again to ensure it still
exists, before trying to access it. This was initially
a band-aid solution to prevent crashing when accessing
channels that no longer existed.

This tames that back a little bit by not doing that
for DAHDI pseudo channels.

Fixes #123