rpt_telemetry: Fix telemetry serialization.
Telemetry was not previously serialized properly, which
at the very least led to lots of these kinds of warnings,
which were added in and around commit
4873199d53acc6f5f5e8e3c83969d3dfbb88594a:

active_telem is NULL, telemetry messages may not be queued properly!
Setting active_telem NULL from (nil), but mytele was 0x7f8bf0107040?

At the time, it was known that a more robust fix for this would be needed;
that was just a temporary fix to ensure the behavior was mostly correct
and telemetry was not queued on top of each other. This is that proper fix.

This is now fixed by using a mutex, rather than just using flags
to keep track of this kind of stuff. As a result, these warning
messages no longer appear. The warning messages are retained since
it is assumed something is wrong if those messages surface.

This has been tested on Murdoch working for 12+ hours. Generally
the warnings surface pretty quickly on a busy system.

Fixes #57
Fixes #124