app_rpt: Eliminate forced telemetry serialization.
This partially reverts commit 0a9e1fee8164dc6b63118dcefbd88f0df1c9a042
and instead eliminates the warnings that those were trying to solve.

An issue exposed by the proposed fix in the aforementioned commit
was that courtesy phones on link unkey were doubled up. There is logic
to avoid this in rpt_telemetry.c, but apparently the logic was now failing
with the telemetry locking and so two courtesy tones were indeed queued.

Further discussion confirmed that telemetry had actually been working
properly in the previous months where the warnings were displaying,
suggesting that the warnings themselves were catching a problem that
wasn't really a problem. As a result, the warnings are now removed,
and the additional locking is removed to restore the correct behavior.

Fixes #128