configs: Cleanup config file extensions. Resolves #167 (#168)
The correct file extension for sample files is conf.sample. It seems the format for these
is a more fully documented file where most lines are commented out. This means that
the rpt_http_registrations.conf file and its sample file appear to have been reversed.
This switches those. It also fixes the echolink file extension from .sample to .conf.

Resolves #167