chan_echolink: Correct locking for el_db_lock, update debug levels (#173)
Changed two routines to properly lock el_db_lock.  This will insure that the data that was returned was used before releasing the lock.  One routine was missing the lock entirely.

Changed the ast_debug levels in several areas.  The main change was to group the internal database update debug level to 5.

Small changes to the debug messages to clarify the data being returned.