chan_echolink: Changes for asterisk standards
This commit updates chan_echolink to conform to the asterisk coding standards. The biggest change was using opening and closing braces on if, do, and while statements. All routines were evaluated and const was added as appropriate. Initialization of some function variables at declaration were removed when the variable was actually assigned to first later in the code. Casts to and from void were removed. New variables were added to some routines so that multiple casts would not be needed later in the code.

Some function declarations were changed to clarify the variable names. Several functions used single letter variables. They now have a more meaningful name. el_node's structure was changed to declare el_pvt as pvt instead of p.

Three logic changes were made: is_rtcp_bye and is_rtcp_sdes now exit their internal loops when they find the sdes packet. There was no need to continue the loop. The code to open and close /dev/null was moved to __mythread_exit.

The logging and debugging statements were reworded to make a complete sentence and/or clarify the data.

mythread_exit was removed after discussion.  It is not needed.

This closes #189