app_rpt / chan_usrp: Bug fixes
This pull request addresses the following issues:

app_rpt: Should recognize an empty string for the scheduler = entry.  #193

app_rpt: The rpt() and rpt_exec() routines depend on ast_waitfor_nandfds for loop timing.  On slower systems, the internal timers do not run reliably.  In my case, they did not run at all.  The timing for these functions have been upgraded to use an actual time calculation.  This makes the internal timers calculate properly in all cases. #195

app_rpt: The rpt dump cli command was not working.  This is corrected.  #192

chan_usrp:  Should unlock the channel when it is created.  #194

chan_usrp: The usrp show cli command  was not working.  It was being unregistered instead of registered.  #194

chan_usrp:  Crashed when receiving audio due to a frame not being initialized.  This has been corrected #198

These changes close #192, #193, #194, #195, and #198