app_rpt: Is not processing GPIOx:x events properly (#211)
app_rpt is not processing GPIOx:x events properly.  Several changes were made to recognize the format used in the event section.  A change was also made in the event routine to properly pass the GPIOx:x information for use by the cop function processing routine.

This has been tested on my system and it works properly.