app_rpt: Outstream resource temporarily unavailable (#223)
During internet outages, while using outstream, the log files quickly fill with the message "resource temporarily unavailable".  In some cases, the system never recovers from the failure.  Resulting in megabytes of log messages.

This change adds a timer to monitor the outstream writes.  The error message is only reported one the first occurrence.  After 60 seconds, if the internet outage has not recovered or the external process has not recovered, app_rpt will kill the external process.  app_rpt will then attempt to start the external process again.  In the event that we do successfully recover, a message is written to the log to indicate a successful recovery.

This closes #212.