res_usbradio: Move user data to safe area in EEPROM
res_usbradio is writing user data over the manufacturing data in the EEPROM for the CM119A
and CM119B.  Memory positions 0 to 50 are reserved for manufacturer specific data.

This change stores the user data in the EEPROM starting at memory position 51 instead of 6.
The internal EEPROM buffer was reduced from 64 unsigned shorts to 13 unsigned shorts.  There
is no need for the extra memory to be allocated since we cannot access or use the manufacturer
data.

All routines in res_usbradio have been documented.

uridiag in the ASL2 repository is being updated to address the memory overwrite.

This addresses one of the issues reported in #227.