chan_simpleusb, chan_usbradio: Reduce frequency of device string not found messages
This changes chan_simpleusb and chan_usbradio to report the device string not found message once.

If the device is found later, the error is cleared.  If the error occurs later, it will be reported again.

This happens when the USB is not plugged in or it fails internally.

This closes #229