chan_simpleusb, chan_usbradio - Refactor functions to res_usbradio
This completes the refactoring of common functions in
chan_simpleusb and chan_usbradio into res_usbradio.

Additional code cleanup was done in chan_simpleusb,
chan_usbradio and res_usbradio.  I added const as needed
to some routines.  I removed unneeded includes in
res_usbradio.

During testing, I discovered a seg fault in chan_usbradio.
This was a result of code that was moved during the
last code cleanup.  The code was restored to the proper
location and the driver now loads and runs.