chan_simpleusb, chan_usbradio:  Correct inconsistent active device message
chan_simpleusb and chan_usbradio printed inconsistent messages
when using the susb tune commands and the associated usb device was
not found.

The message now reads "Device xxxx is selected, the associated USB
device string yyyy was not found".

This will help users understand the nature of the problem.

This closes #258.