rpt_manager: Prevent NULL dereference of channel.
If app_rpt is being unloaded, we can potentially
be trying to use rpt_vars[i].rxchannel when it
is made NULL by another thread, as this part of
the code is not protected by the node lock. This
can occasionally result in a segfault due to a
NULL dereference (trying to access the varshead
of NULL).

This is fixed easily by saving a copy of the channel
pointer and using that so we know it won't change
while we're using it. This is safe because we already
ref the channel beforehand, so we know the channel
itself can't disappear while we're using it.

Fixes #268