CI: Fix missing ALSA dependency in build.
Asterisk 21 removed chan_alsa, and with it the ALSA lib
detection that existed in the build system. This has
prevented the USB modules from building on Asterisk 21,
though they work fine on Asterisk <= 20.

PhreakScript has been modified to support readding ALSA
lib detection (not chan_alsa itself), which now allows
us to build the USB modules that depend on ALSA.

This avoids needing to remove menuselect.makeopts and
recompile everything else as well, so this should cut
build times in half and do a more complete build.