rpt_install.sh: Download tarball of sounds to avoid SVN.
Previously, we had been using SVN to download the sounds
from GitHub. However, GitHub has removed support for SVN
as of a couple weeks ago, which has broken this workflow.

See: https://github.blog/2023-01-20-sunsetting-subversion-support/

SVN was used to avoid cloning the entire repo; however,
as the sounds aren't changing frequently, we now download
a tarball that has been added to the website and extract
them into the correct directory.