chan_usbradio / chan_simpleusb: Changes for queue size
chan_usbradio was reporting queue size overflows.  The
driver has been updated with more error reporting code.
Some buffering code was removed as it served no purpose.
The default queue size was changed from 2 to 20.  This
moves from a 40 millisecond output buffer to a 400
millisecond buffer.

chan_simpleusb was updated with the same error reporting
code.  The queue size in this driver was 100 milliseconds.
Tests indiate this is adequate.

This closes #272.