chan_usbradio: Change to prevent buffer overflow
This changes chan_usbradio to only send audio to the sound
device when the PTT is active.  This helps address a
problem with the internal timing difference causing the
driver to overflow the sound device's internal buffer.

This makes chan_usbradio work like chan_simpleusb.  chan_simpleusb
only sends audio when PTT is active.