chan_usbradio and chan_simpleusb: Add auto assign device string when blank
chan_usbradio and chan_simpleusb will now automatically assign a devicestring, 'devstr', when it is blank.

This allows the default configuration files to be published with devstr left blank. The default hardcoded value for devstr may not be the correct value on the end user's computer.

This will allow users to start up with minimal configuration changes. When the user saves the tune settings, the automatically assigned devstr will be written to the configuration file.

This closes #284.