chan_echolink: Address module load decline due to dns
This updates chan_echolink to address a problem where the
module declines to load because it cannot find perform
a DNS lookup for the APRS server.

This should not stop the driver from starting.  The driver
will now attempt to perform the DNS lookup when it needs
to send an APRS packet.  That occurs every 10 minutes.

This closes #309