Template-ize simpleusb.conf and usbradio.conf
In the past, configuring more than one SimpleUSB (or USBRadio) node on a system
required duplication of large chunks of settings in the "simpleusb.conf" (and
"usbradio.conf") file.  Managing these configuraitons in the menu(s) was also
messy.  Updating the files to use templates greatly simplifies our support for
multiple nodes.