[rpt] HTTP registrations incorrectly showed as "Registered"
When attempting to register a node w/HTTP and when not providing a valid node
and password the status was incorrectly being reported as "Registered".  This
was because the text of the response "message" was not being checked.

One could argue that this would be better handled by having the registration
server respond to the request with an non "2XX" code.  Unfortunately, this is
how the registration server was setup years ago.  Changing (correcting) the
behavior would require server changes and possible changes to this code.  We
would also need to ensure that any older (already deployed) systems would
continue to function correctly.

For now, we will check the HTTP response text (as was done in the past).