Update directory and file permissions
The idea here is that we keep the "archivedir" directory permissions at 700 but
create the subdirecties as 755 and the [leaf] files as 644.  No more errors due
to the incorrect directory permissions.  Access to the archive starts off locked
down but if the node owner opens the "archivedir" permissions then the rest of
the content becomes available without having to perform extra chmod's.