Correct incorrect "archivedir" filesystem permissions (#342)
* Correct incorrect "archivedir" filesystem permissions

Newly created directories should be created with the eXecute bit set (e.g. 0700).
Found and fixed a few calls to mkdir() that were setting mode 0600 (incorrect)
instead of the mode 0700 (expected).

Also corrected two log messages that were correctly/fully reporting the specific
filenames that could not be openned.

* Update directory and file permissions

The idea here is that we keep the "archivedir" directory permissions at 700 but
create the subdirecties as 755 and the [leaf] files as 644.  No more errors due
to the incorrect directory permissions.  Access to the archive starts off locked
down but if the node owner opens the "archivedir" permissions then the rest of
the content becomes available without having to perform extra chmod's.

* Open up sub-directory group write permissions

Scripts processing the files added to the "archivedir" directory may actually
want to remove files that have been processed.  Extend the "group" privs to allow
removing files.

Note: ownership of the "archivedir" dirs/files will be "asterisk:asterisk"