chan_tlb: Code updates and documentation added (#343)
* Update chan_tlb to address a channel deadlock

This updates chan_tlb with address a deadlock on the channel.  The channel was not unlocked when it was created.

This closes #339

* Update chan_tlb to address a channel deadlock

This updates chan_tlb with address a deadlock on the channel. The channel was not unlocked when it was created.

This closes #339

* chan_tlb: Code updates and documentation

The updates chan_tlb with coding standards and documentation.

Code was reordered, and unnecessary variables were removed.

---------

Co-authored-by: Allan Nathanson <42244061+Allan-N@users.noreply.github.com>