Fix issue where bad registration info leads to NULL dereference crash
When you start off with a "rpt_http_registrations.conf" containing an error and
attempt to register then you may get back an unexpected response.  If you don't
check for the unexpected then you can end up with a crash.  This change adds a
missing check for the unexpected (lack of "data").