app_rpt: Update to use DNS SRV record (#361)
* app_rpt: Update to use DNS SRV record

This updates app_rpt to use DNS SRV records when looking up a node by DNS.

The original code used the TXT record which was labeled for debugging purposes.  The original TXT lookup code remains and is switched off for now.

This closes #360

* app_rpt: Update to use DNS SRV record

This updates app_rpt to use DNS SRV records when looking up a node by DNS.

The original code used the TXT record which was labeled for debugging purposes. The original TXT lookup code remains and is switched off for now.

This closes #360

* app_rpt: Update to use DNS SRV record

This updates app_rpt to use DNS SRV records when looking up a node by DNS.

The original code used the TXT record which was labeled for debugging purposes. The original TXT lookup code remains and is switched off for now.

This closes #360

* app_rpt: Update to use DNS SRV record

This updates app_rpt to use DNS SRV records when looking up a node by DNS.

The original code used the TXT record which was labeled for debugging purposes. The original TXT lookup code remains and is switched off for now.

This closes #360