app_rpt: Correct issues with the autopatch
This addresses a deadlock and problems with joining the voxchannel conference room with the autopatch.

dahdi_conf_fd_confno now uses ioctl DAHDI_GETCONF instead of DAHDI_CHANNO.  I determined
that DAHDI_CHANNO returns the channel number, not the conference number.  DAHDI_CHANNO returned
the conference number in some cases, but not in all of them.  DAHDI_GETCONF works properly
in all cases.

The voxchannel could not join the conference using the flag DAHDI_CONF_MONITOR.  ioctl would
return EINVAL.  After testing I changed it to RPT_CONF_CONF | RPT_CONF_LISTENER.  This allowed
the voxchannel to join the conference.  I did not see any side effects, so this needs to be
carefully tested.

This closes #257.