app_rpt: Correct 5 digit node lookup from 4 digit node
This updates app_rpt to address a problem with looking up a 5 or 6
digit node from a 4 digit node.  The 4 digit node could not
lookup 5 or 6 digit nodes when it did not use an external nodes
file.  This would happen if you are not using the node updater and
depending on DNS only.

The problem was associated with the internal 'longestnode' variable.
We did not have an external file to determine the longest node number.
In this case, the longest node was set to 4.

A new define, 'MIN_NODE_LENGTH' has been added with a value of 6.  The
routines that determine the longest nodes will now MAX the value
it determines with MIN_NODE_LENGTH.