app_rpt: Correct 5 digit node lookup from 4 digit node (#372)
* app_rpt: Correct 5 digit node lookup from 4 digit node

This updates app_rpt to address a problem with looking up a 5 or 6
digit node from a 4 digit node.  The 4 digit node could not
lookup 5 or 6 digit nodes when it did not use an external nodes
file.  This would happen if you are not using the node updater and
depending on DNS only.

The problem was associated with the internal 'longestnode' variable.
We did not have an external file to determine the longest node number.
In this case, the longest node was set to 4.

A new define, 'MIN_NODE_LENGTH' has been added with a value of 6.  The
routines that determine the longest nodes will now MAX the value
it determines with MIN_NODE_LENGTH.

* app_rpt: Correct 5 digit node lookup from 4 digit node

This updates app_rpt to address a problem with looking up a 5 or 6 digit node from a 4 digit node. The 4 digit node could not lookup 5 or 6 digit nodes when it did not use an external nodes file. This would happen if you are not using the node updater and depending on DNS only.

The problem was associated with the internal 'longestnode' variable. We did not have an external file to determine the longest node number. In this case, the longest node was set to 4.

A new rpt.conf [general] item, maximum_node_length, has been added with a default value of 6. The routines that determine the longest nodes will now MAX the value it determines with maximum_node_length.  The configurable item cannot be less than 4.

By default, this configurable item will not be in rpt_conf.  It can be used in the future if we need to support longer node numbers.

Without this change, the problem can be remediated with existing installs, that don't use the node updater, by adding the following to rpt.conf [nodes].

999999 = radio@127.0.0.1:4569/999999,NONE

* app_rpt: Correct 5 digit node lookup from 4 digit node

This updates app_rpt to address a problem with looking up a 5 or 6 digit node from a 4 digit node. The 4 digit node could not lookup 5 or 6 digit nodes when it did not use an external nodes file. This would happen if you are not using the node updater and depending on DNS only.

The problem was associated with the internal 'longestnode' variable. We did not have an external file to determine the longest node number. In this case, the longest node was set to 4.

A new rpt.conf [general] item, maximum_node_length, has been added with a default value of 6. The routines that determine the longest nodes will now MAX the value it determines with maximum_node_length. The configurable item cannot be less than 4 or more then 63.

By default, this configurable item will not be in rpt_conf. It can be used in the future if we need to support longer node numbers.

Without this change, the problem can be remediated with existing installs, that don't use the node updater, by adding the following to rpt.conf [nodes].

999999 = radio@127.0.0.1:4569/999999,NONE

* app_rpt: Correct 5 digit node lookup from 4 digit node

This updates app_rpt to address a problem with looking up a 5 or 6 digit node from a 4 digit node. The 4 digit node could not lookup 5 or 6 digit nodes when it did not use an external nodes file. This would happen if you are not using the node updater and depending on DNS only.

The problem was associated with the internal 'longestnode' variable. We did not have an external file to determine the longest node number. In this case, the longest node was set to 4.

A new rpt.conf [general] item, max_dns_node_length, has been added with a default value of 6. The routines that determine the longest nodes will now MAX the value it determines with rpt_max_dns_node_length. The configurable item cannot be less than 4 or more then 63.

By default, this configurable item will not be in rpt_conf. It can be used in the future if we need to support longer node numbers.

Without this change, the problem can be remediated with existing installs, that don't use the node updater, by adding the following to rpt.conf [nodes].

999999 = radio@127.0.0.1:4569/999999,NONE

* app_rpt: Correct 5 digit node lookup from 4 digit node

This updates app_rpt to address a problem with looking up a 5 or 6 digit node from a 4 digit node. The 4 digit node could not lookup 5 or 6 digit nodes when it did not use an external nodes file. This would happen if you are not using the node updater and depending on DNS only.

The problem was associated with the internal 'longestnode' variable. We did not have an external file to determine the longest node number. In this case, the longest node was set to 4.

A new rpt.conf [general] item, max_dns_node_length, has been added with a default value of 6. The routines that determine the longest nodes will now MAX the value it determines with rpt_max_dns_node_length. The configurable item cannot be less than 4 or more then 63.

By default, this configurable item will not be in rpt_conf. It can be used in the future if we need to support longer node numbers.

Additional level 4 debugging was added to the DNS lookup routine.

Without this change, the problem can be remediated with existing installs, that don't use the node updater, by adding the following to rpt.conf [nodes].

999999 = radio@127.0.0.1:4569/999999,NONE