app_rpt: Correct crash when using rpt page
This updates app_rpt to address a crash that occurred when using the rpt page command.

ast_sendtext should not require locking according to the documentation, but that does
not seem to be the case.  In some instances, the main processing loop had control of
the channel.  ast_sendtext would cause an assert abort in channel.c.

This was shown by the message ast_sendtext_data: Thread LWP 88692 is blocking 'Radio/usb_2231',
already blocked by thread LWP 88688 in procedure ast_waitfor_nandfds.

Adding rpt_mutex_lock around ast_sendtext resolved the problem.

Corrected an if statement in the same routine that was already flagged in the code
as a bug.

Applied coding standards to the routine.

This closes #375.