app_rpt: Correct VOX operation for autopatch
This change restores the operation of the autopatch VOX function.

Changes were made in #363 to get the autopatch working. Part of those changes did not correctly address the operation of the vox channel. This change removes some of those changes and restores the operation of the vox channel.

For the vox channel to work properly, this change along with changes to the DAHDI driver are required. The changes to the DAHDI driver are located at https://github.com/KB4MDD/dahdi-linux

We will be attempting to get these changes into the main DAHDI channel.

This closes #388