Fix LastKeyedTime/LastTxKeyedTime info in RPT events (#406)
The rpt_manager_trigger() function uses ctime(3) to add formatted date/time
strings for LastKeyedTime and LastTxKeyedTime.  Unfortunately, the returned
strings include trailing '\n' characters that added empty lines to the output.
This presents an issue as an empty line typically indicates the end of the AMI
output.

This change removes the extra (unexpected) '\n' characters.