chan_simpleusb, chan_usbradio: Add parameter to make scaling/clipping optional, fix issue #399 (#418)
The SimpleUSB and USBRadio channel drivers included code that attempted to match levels to the original CM108 IC (a chip has now been out of production for over 10 years). The code was scaling audio to 109.375% with possible clipping.  Any adjustments for CM1xxx gain differences should be really be made in the mixer settings, not in the audio stream.

This change adds a new `legacyaudioscaling` variable to the drivers that control whether the scaling / clipping code present in earlier ASL releases is enabled.  By default, this variable is set to "yes" to ensure that the behavior of existing configurations remains the same.  The `simpleusb.conf` and `usbradio.conf` files, used with new installs, will include `legacyaudioscaling = no` to avoid the potential clipping behavior.