app_rpt:  Channel initialization for PTT
app_rpt is not waiting long enough in rpt_setup_channels for chan_simpleusb to
initialize the usb device before asserting PTT on/off.

This change adds a 250ms delay before asserting PTT on and another 50ms delay
before asserting PTT off.

I believe the intention of the original code was to insure that the PTT was
toggled immediately after the channel driver was created by app_rpt.

Testing by the effected users is needed. This may close #421 and #424.